Switch to 2-stage scan with variable block width. Correct conditioning error.
An error in the conditioning structure of the approximate E-step has been
corrected. The region considered in likelihood calculations was formerly padded
only by +/- w, where w is the template's half-width. It is now (correctly)
padded by 2*w, as this is the distance required to include all nucleosomes that
could contribute reads to the region under consideration.

With the simpler, variable block-width scan, the M-step is now performed on
every outer iteration. This massively accelerated convergence and improves
results.