Major revision throughout. Replace block scan, fix conditioning, etc.
The block scan algorithm has been simplified and modified to avoid low coverage
at the ends of the chromosome. This uses variable-width blocks and necessitates
a bit more handling on the workers.

Errors in conditioning and the proposal density calculation have been corrected
in the worker MH step. The region considered in likelihood calculations was
formerly padded only by +/- w, where w is the template's half-width. It is now
(correctly) padded by 2*w, as this is the distance required to include all
nucleosomes that could contribute reads to the region under consideration.

A minor error in the sigmasq draw has been corrected as well.

Options to initialize mu, sigmasq, and theta from EM output are also now
available and tested.

The MH step on theta is not working well. HMC may be needed.