Retain acceptance rate. Increase HMC step size. Use newest pickle protocol.
HMC step size increased from 0.01 to 0.1. This increased effective sample sizes
by a factor of 5-10 and reduced acceptance rates to around 80% for the HMC step
with a block width of 400bp.

All cPickle calls have been changed to use the latest protocol. This is binary
and much more efficient than the default (test-based) protocol 0.