Add fix for blocks shorter than conditioning buffer.
To enable distributed deconvolution, we divide each chromosome into blocks
separated by 2 * w, w = floor(length(template) / 2). Conditioning on buffers of
at this this size makes the blocks conditionally independent in the posterior.
There was a subtle bug in our handling of blocks with length <= w.

Algorithmically, blocks are defined as adjacent, non-overlapping segments of the
chromosome. These include buffers of length w for conditioning. It is each
worker's reponsibility to ensure appropriate conditioning given such a buffer.
For example, with a block of length 1000 and w = 50, each worker updates only
entries 50:950.

In the original implementation, it was posible to have blocks of length < w at
the end of a chromosome. This led to negative dimension errors in the worker
functions, reflecting the impossiblity of conditioning correctly.

This fix removes such blocks before running the MCMC iterations. This can
slightly reduce update coverage at the end of the chromosome, but all entries of
theta are still updated at least once per outer iteration.